add sqlite support

+++ database/migrations/2015_03_19_124805_add_src_reference_column_to_translations.php
@@ -14,7 +14,7 @@ class AddSrcReferenceColumnToTranslations extends Migration
     public
     function up()
     {
-        Schema::table('ltm_translations', function (Blueprint $table)
+        Schema::connection('sqlite')->table('ltm_translations', function (Blueprint $table)
         {
             $table->string('source', 256)->nullable();
             $table->unique(['locale','group','key'], 'ixk_ltm_translations_locale_group_key');
@@ -29,10 +29,12 @@ class AddSrcReferenceColumnToTranslations extends Migration
     public
     function down()
     {
-        Schema::table('ltm_translations', function (Blueprint $table)
+        Schema::connection('sqlite')->table('ltm_translations', function (Blueprint $table)
         {
-            $table->dropColumn('source');
-            $table->dropIndex('ixk_ltm_translations_locale_group_key');
+            if (Schema::connection('sqlite')->hasColumn('ltm_translations', 'source')) {
+                $table->dropColumn('source');
+            }
+            //$table->dropIndex('ixk_ltm_translations_locale_group_key');
         });
     }
 }
